requirements.txt
langchain
langchain-ollama
python-dotenv
langchain_community
langchain_core
streamlit